requirements.txt
asgiref==3.6.0
astroid==2.13.3
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
Django==4.1.2
django-render-block==0.9.2
docx2txt==0.8
gTTS==2.3.1
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pylint==2.15.10
PyPDF4==1.27.0
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
